proper labels for subgroups, display hierarchy lines, moved logic to calculator for better testability
