+++ webapp/src/main/webapp/src/styles.less
@@ -367,8 +367,18 @@ sb-button-typeahead .input-group-btn > .btn,
   max-width: 330px;
 }
 
+.wrapping-max-width-xl {
+  white-space: normal;
+  text-align: left;
+  max-width: 500px;
+}
+
 .no-wrap {
   white-space:nowrap;
+}
+
+.subgroup-toggle {
+  margin-left: 5px;
 }
 
 /**
